Deal with version problem on aird file: New minor versions lead to not being able to open them on an older version, The file's version 15.4.3 is only availble from single package upgrades, while sirius can be updated to 15.4.1 using emf's general update site https://download.eclipse.org/modeling/emf/emf/builds/index.html
Although only aird really changed, we constructed it via two commits: revert and re-do that are now squashed. Closes #3

Revert "added attributes feltTrustImmediately and feltTrustAfterwards to class"
as far as it is about the model since the newer editor version caused trouble with opening the file (see https://github.com/keml-group/keml/issues/3)
We need to downgrade to a version that is available in a normal bundle.

This reverts commit 64682083b6deac46c3b3c7773f5f2891d0d64f51.

Add felt Trust immediately and afterwards on information again (after revert), now in the newest version that is available as whole modeling bundle, 15.4.1